fix(parser): use usize for tilde overflow check to match brush's parse type
decision(parser): brush-parser's TildeExpr fields are usize, not u64;
identical on every 64-bit release target but usize is the correct
threshold by construction rather than by coincidence of target width.

learned(parser): the dir-stack comment mis-attributed both `~+<n>` and
`~-<n>` to distinct brush rules; `~+<n>` actually reaches the same rule
as plain `~<n>` (only the `+` is optional there), only `~-<n>` is a
separate rule.

+++ src/parser.rs
@@ -1227,7 +1227,7 @@ fn convert_brace_members(
 }
 
 /// When `text` starts with a `~` directly followed by an optional `+`/`-`
-/// sign and a decimal digit run that overflows `u64`, returns the
+/// sign and a decimal digit run that overflows `usize`, returns the
 /// `(tilde_run, remainder)` split at the end of that digit run. This is
 /// the exact shape that panics brush-parser 0.4.0's tilde-expression
 /// parser: the unsigned/`+`-prefixed dir-stack form (`word.rs:909`,
@@ -1251,7 +1251,7 @@ fn split_overflowing_leading_tilde(text: &str) -> Option<(&str, &str)> {
     }
     let (digits, remainder) = digits_start.split_at(digit_len);
     digits
-        .parse::<u64>()
+        .parse::<usize>()
         .is_err()
         .then(|| (&text[..1 + sign_len + digit_len], remainder))
 }
